Separate out schema errors into their own file
