+++ schema/schema_formats.go
@@ -102,83 +102,6 @@ type FormatSchemaInstance struct {
 	Format      string `json:"format"` // value set from parent FormatSchema's `CanonicalName`
 }
 
-// Format/schema error types
-type UnsupportedFormatError struct {
-	Type      string
-	Message   string
-	InputFile string
-	Format    string
-	Version   string
-	Variant   string
-	Command   string
-	Flags     string
-}
-
-type UnsupportedSchemaError struct {
-	UnsupportedFormatError
-}
-
-func NewUnsupportedSchemaError(m string, format string, version string, variant string) *UnsupportedSchemaError {
-	var err = new(UnsupportedSchemaError)
-	err.Type = ERR_TYPE_UNSUPPORTED_SCHEMA
-	err.Message = m
-	err.Format = format
-	err.Version = version
-	err.Variant = variant
-	return err
-}
-
-func NewUnsupportedFormatError(msg string, f string, fmt string, cmd string, flags string) *UnsupportedFormatError {
-	var err = new(UnsupportedFormatError)
-	err.Type = ERR_TYPE_UNSUPPORTED_FORMAT
-	err.Message = msg
-	err.InputFile = f
-	err.Format = fmt
-	err.Command = cmd
-	err.Flags = flags
-	return err
-}
-
-func NewUnsupportedFormatForCommandError(f string, fmt string, cmd string, flags string) *UnsupportedFormatError {
-	var err = new(UnsupportedFormatError)
-	err.Type = ERR_TYPE_UNSUPPORTED_FORMAT
-	err.Message = MSG_FORMAT_UNSUPPORTED_COMMAND
-	err.InputFile = f
-	err.Format = fmt
-	err.Command = cmd
-	err.Flags = flags
-	return err
-}
-
-func NewUnknownFormatError(f string) *UnsupportedFormatError {
-	var err = new(UnsupportedFormatError)
-	err.Type = ERR_TYPE_UNSUPPORTED_FORMAT
-	err.Message = MSG_FORMAT_UNSUPPORTED_UNKNOWN
-	err.InputFile = f
-	return err
-}
-
-func (err UnsupportedFormatError) Error() string {
-	baseMessage := fmt.Sprintf("%s: %s (`%s`)", err.Type, err.Message, err.InputFile)
-	if err.Format != "" {
-		return fmt.Sprintf("%s: format: `%s`, command: `%s`, flags: `%s`",
-			baseMessage,
-			err.Format,
-			err.Command,
-			err.Flags)
-	}
-	return baseMessage
-}
-
-func (err UnsupportedSchemaError) Error() string {
-	return fmt.Sprintf("%s: %s: Schema Format: `%s`, Version: `%s`, Variant: `%s` ",
-		err.Type,
-		err.Message,
-		err.Format,
-		err.Version,
-		err.Variant)
-}
-
 func (config *BOMFormatAndSchemaConfig) Reset() {
 	config.Formats = nil
 }
